requirements.txt
pybullet==3.2.6
numpy==1.26.4
requests==2.32.3
pydantic==2.8.2
python-dotenv==1.0.1
pytest==8.3.2
